+++ NEWS.md
@@ -17,7 +17,7 @@
 * `tableby()`: Escaped `%` for `text="latex"`. (#258)
 
 * `tableby()` / `paired()`: Added support for "min", "max", "sd", "mean", and "var" (#259) and
-  "gmean", "gsd", "gmeansd", "gmeanCI" (#260) and "Npct" (#263).
+  "gmean", "gsd", "gmeansd", "gmeanCI" (#260) and "Npct" (#263) and "sum" (#281).
   
 * `tableby()` / `paired()`: Added a more informative error message when no summary statistic is computed. (#273)